Move uniforms from p5.Renderer to p5.Shader
